requirements.txt
asgiref==3.7.2
Django==4.2.3
sqlparse==0.4.4
psycopg2-binary==2.9.6
djangorestframework==3.14.0
Pillow==10.0.0